feat(http): move trace package
