+++ pkg/http/client.go
@@ -4,6 +4,8 @@ import (
 	"net/http"
 
 	pkgZap "github.com/alexfalkowski/go-service/pkg/http/logger/zap"
+	"github.com/alexfalkowski/go-service/pkg/http/meta"
+	"github.com/alexfalkowski/go-service/pkg/http/trace/opentracing"
 	"go.uber.org/zap"
 )
 
@@ -11,8 +13,8 @@ import (
 func NewRoundTripper(logger *zap.Logger) http.RoundTripper {
 	hrt := http.DefaultTransport
 	hrt = pkgZap.NewRoundTripper(logger, hrt)
-	hrt = &traceRoundTripper{RoundTripper: hrt}
-	hrt = &metaRoundTripper{RoundTripper: hrt}
+	hrt = opentracing.NewRoundTripper(hrt)
+	hrt = meta.NewRoundTripper(hrt)
 
 	return hrt
 }
