--- pkg/http/http.go
@@ -1,14 +1,0 @@
-package http
-
-const (
-	httpRequest         = "http.request"
-	httpResponse        = "http.response"
-	httpURL             = "http.url"
-	httpMethod          = "http.method"
-	httpDuration        = "http.duration_ms"
-	httpStartTime       = "http.start_time"
-	httpRequestDeadline = "http.request.deadline"
-	httpStatusCode      = "http.status_code"
-	component           = "component"
-	httpComponent       = "http"
-)
