+++ pkg/http/trace/opentracing/encode.go
@@ -1,4 +1,4 @@
-package http
+package opentracing
 
 import (
 	"bytes"
